+++ projects/js-packages/idc/components/idc-screen/index.jsx
@@ -19,7 +19,7 @@ import IDCScreenVisual from './visual';
 const IDCScreen = props => {
 	const {
 		logo,
-		customContent,
+		customContent = {},
 		wpcomHomeUrl,
 		currentUrl,
 		apiNonce,
@@ -112,10 +112,6 @@ IDCScreen.propTypes = {
 	isAdmin: PropTypes.bool.isRequired,
 	/** If potentially dynamic HTTP_HOST usage was detected for site URLs in wp-config which can lead to a JP IDC. */
 	possibleDynamicSiteUrlDetected: PropTypes.bool,
-};
-
-IDCScreen.defaultProps = {
-	customContent: {},
 };
 
 export default IDCScreen;