+++ projects/js-packages/publicize-components/src/social-store/actions/connection-data.js
@@ -2,23 +2,39 @@ import { globalNoticesStore } from '@automattic/jetpack-components';
 import apiFetch from '@wordpress/api-fetch';
 import { dispatch as coreDispatch } from '@wordpress/data';
 import { store as editorStore } from '@wordpress/editor';
-import { __ } from '@wordpress/i18n';
+import { __, sprintf } from '@wordpress/i18n';
 import {
+	ADD_CONNECTION,
+	CREATING_CONNECTION,
 	DELETE_CONNECTION,
 	DELETING_CONNECTION,
 	SET_CONNECTIONS,
 	TOGGLE_CONNECTION,
+	UPDATE_CONNECTION,
+	UPDATING_CONNECTION,
 } from './constants';
 
 /**
  * Set connections list
- * @param {Array} connections - list of connections
+ * @param {Array<import('../types').Connection>} connections - list of connections
  * @returns {object} - an action object.
  */
 export function setConnections( connections ) {
 	return {
 		type: SET_CONNECTIONS,
 		connections,
+	};
+}
+
+/**
+ * Add connection to the list
+ * @param {import('../types').Connection} connection - connection object
+ * @returns {object} - an action object.
+ */
+export function addConnection( connection ) {
+	return {
+		type: ADD_CONNECTION,
+		connection,
 	};
 }
 
@@ -162,6 +178,19 @@ export function deletingConnection( connectionId, deleting = true ) {
 }
 
 /**
+ * Whether a connection is being created.
+ *
+ * @param {boolean} creating - Whether the connection is being creating.
+ * @returns {object} Creating connection action.
+ */
+export function creatingConnection( creating = true ) {
+	return {
+		type: CREATING_CONNECTION,
+		creating,
+	};
+}
+
+/**
  * Deletes a connection by disconnecting it.
  *
  * @param {object} args - Arguments.
@@ -199,6 +228,136 @@ export function deleteConnectionById( { connectionId, showSuccessNotice = true }
 			createErrorNotice( message, { type: 'snackbar', isDismissible: true } );
 		} finally {
 			dispatch( deletingConnection( connectionId, false ) );
+		}
+	};
+}
+
+/**
+ * Creates a connection.
+ *
+ * @param {Record<string, any>} data - The data for API call.
+ * @returns {void}
+ */
+export function createConnection( data ) {
+	return async function ( { dispatch } ) {
+		const { createErrorNotice, createSuccessNotice } = coreDispatch( globalNoticesStore );
+
+		try {
+			const path = `/jetpack/v4/social/connections/`;
+
+			dispatch( creatingConnection() );
+
+			const connection = await apiFetch( { method: 'POST', path, data } );
+
+			if ( connection ) {
+				dispatch(
+					addConnection( {
+						...connection,
+						// TODO fix this messy data structure
+						connection_id: connection.ID.toString(),
+						display_name: connection.external_display,
+						service_name: connection.service,
+						external_id: connection.external_ID,
+						profile_link: connection.external_profile_URL,
+						profile_picture: connection.external_profile_picture,
+						can_disconnect: true,
+					} )
+				);
+
+				createSuccessNotice(
+					sprintf(
+						/* translators: %s is the name of the social media platform e.g. "Facebook" */
+						__( '%s account connected successfully.', 'jetpack' ),
+						connection.label
+					),
+					{
+						type: 'snackbar',
+						isDismissible: true,
+					}
+				);
+			}
+		} catch ( error ) {
+			let message = __( 'Error connecting account.', 'jetpack' );
+
+			if ( typeof error === 'object' && 'message' in error && error.message ) {
+				message = `${ message } ${ error.message }`;
+			}
+
+			createErrorNotice( message, { type: 'snackbar', isDismissible: true } );
+		} finally {
+			dispatch( creatingConnection( false ) );
+		}
+	};
+}
+
+/**
+ * Updates a connection.
+ *
+ * @param {string} connectionId - Connection ID to update.
+ * @param {Record<string, any>} data - The data.
+ *
+ * @returns {object} Delete connection action.
+ */
+export function updateConnection( connectionId, data ) {
+	return {
+		type: UPDATE_CONNECTION,
+		connectionId,
+		data,
+	};
+}
+
+/**
+ * Marks a connection as being updating.
+ *
+ * @param {string} connectionId - Connection ID being updated.
+ * @param {boolean} updating - Whether the connection is being updated.
+ *
+ * @returns {object} Deleting connection action.
+ */
+export function updatingConnection( connectionId, updating = true ) {
+	return {
+		type: UPDATING_CONNECTION,
+		connectionId,
+		updating,
+	};
+}
+
+/**
+ * Updates a connection.
+ *
+ * @param {string} connectionId - Connection ID to update.
+ * @param {Record<string, any>} data - The data for API call.
+ * @returns {void}
+ */
+export function updateConnectionById( connectionId, data ) {
+	return async function ( { dispatch } ) {
+		const { createErrorNotice, createSuccessNotice } = coreDispatch( globalNoticesStore );
+
+		try {
+			const path = `/jetpack/v4/social/connections/${ connectionId }`;
+
+			dispatch( updatingConnection( connectionId ) );
+
+			const connection = await apiFetch( { method: 'POST', path, data } );
+
+			if ( connection ) {
+				dispatch( updateConnection( connectionId, data ) );
+
+				createSuccessNotice( __( 'Account updated successfully.', 'jetpack' ), {
+					type: 'snackbar',
+					isDismissible: true,
+				} );
+			}
+		} catch ( error ) {
+			let message = __( 'Error updating account.', 'jetpack' );
+
+			if ( typeof error === 'object' && 'message' in error && error.message ) {
+				message = `${ message } ${ error.message }`;
+			}
+
+			createErrorNotice( message, { type: 'snackbar', isDismissible: true } );
+		} finally {
+			dispatch( updatingConnection( connectionId, false ) );
 		}
 	};
 }