+++ projects/packages/publicize/src/jetpack-social-settings/class-settings.php
@@ -10,7 +10,6 @@ namespace Automattic\Jetpack\Publicize\Jetpack_Social_Settings;
 use Automattic\Jetpack\Connection\Client;
 use Automattic\Jetpack\Connection\Manager;
 use Automattic\Jetpack\Modules;
-use Automattic\Jetpack\Publicize\Publicize;
 use Automattic\Jetpack\Publicize\Social_Image_Generator\Templates;
 
 /**
@@ -183,13 +182,14 @@ class Settings {
 
 		$settings = $this->get_settings( true );
 
-		$settings['useAdminUiV1'] = Publicize::use_admin_ui_v1();
+		$settings['useAdminUiV1'] = false;
 
 		$settings['is_publicize_enabled'] = false;
 
 		$connection = new Manager();
 
 		if ( ( new Modules() )->is_active( 'publicize' ) && $connection->has_connected_user() ) {
+			$settings['useAdminUiV1']   = $publicize->use_admin_ui_v1();
 			$settings['connectionData'] = array(
 				'connections' => $publicize->get_all_connections_for_user(),
 				'adminUrl'    => esc_url_raw( $publicize->publicize_connections_url( 'jetpack-social-connections-admin-page' ) ),