+++ projects/plugins/jetpack/extensions/blocks/slideshow/slideshow.js
@@ -120,12 +120,13 @@ class Slideshow extends Component {
 			return null;
 		}
 
-		const { autoplay, delay, effect, images } = this.props;
+		const { autoplay, className, delay, effect, images } = this.props;
 		// Note: React omits the data attribute if the value is null, but NOT if it is false.
 		// This is the reason for the unusual logic related to autoplay below.
 		/* eslint-disable jsx-a11y/anchor-is-valid */
 		return (
 			<div
+				className={ className }
 				data-autoplay={ autoplay || null }
 				data-delay={ autoplay ? delay : null }
 				data-effect={ effect }