+++ projects/packages/connection/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.8.2] - 2024-05-16
+### Added
+- Connection: Ensuring direct file access is disabled in class-jetpack-ixr-client.php [#37398]
+
+### Changed
+- Updated package dependencies. [#37379]
+
+## [2.8.1] - 2024-05-14
+### Changed
+- SSO: do not rely on the Jetpack class anymore. [#37153]
+
 ## [2.8.0] - 2024-05-13
 ### Added
 - SSO: Ensuring tooltips are accessible [#37302]
@@ -1058,6 +1069,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Separate the connection library into its own package.
 
+[2.8.2]: https://github.com/Automattic/jetpack-connection/compare/v2.8.1...v2.8.2
+[2.8.1]: https://github.com/Automattic/jetpack-connection/compare/v2.8.0...v2.8.1
 [2.8.0]: https://github.com/Automattic/jetpack-connection/compare/v2.7.7...v2.8.0
 [2.7.7]: https://github.com/Automattic/jetpack-connection/compare/v2.7.6...v2.7.7
 [2.7.6]: https://github.com/Automattic/jetpack-connection/compare/v2.7.5...v2.7.6